+++ src/components/Badge/Badge.style.ts
@@ -1,6 +1,7 @@
 "use client";
 
 import styled from "styled-components";
+import DESIGN_SYSTEM from "@/styles/designSystem";
 
 /**
  * Badge 컴포넌트에 필요한 style props입니다
@@ -13,7 +14,7 @@ export const DotBadgeContainer = styled.div<IBadgeStyle>`
   display: flex;
   justify-content: center;
   align-items: center;
-  padding: ${({ theme }) => theme.gap?.["4xs"]};
+  padding: ${DESIGN_SYSTEM.gap["4xs"]};
 
   width: fit-content;
   height: fit-content;
@@ -22,9 +23,8 @@ export const DotBadgeContainer = styled.div<IBadgeStyle>`
 export const Dot = styled.div`
   width: 4px;
   height: 4px;
-  border-radius: ${({ theme }) => theme.radius?.circle};
-  background-color: ${({ theme }) =>
-    theme.colors?.light["feedback-notification"]};
+  border-radius: ${DESIGN_SYSTEM.radius.circle};
+  background-color: ${({ theme }) => theme.light["feedback-notification"]};
 `;
 
 export const TextBadgeContainer = styled.div<IBadgeStyle>`
@@ -33,15 +33,14 @@ export const TextBadgeContainer = styled.div<IBadgeStyle>`
   min-width: 20px;
   padding: ${(props) =>
     props.$variant === "count"
-      ? `${props.theme.gap?.none} ${props.theme.gap?.["4xs"]}`
-      : props.theme.gap?.none};
+      ? `${DESIGN_SYSTEM.gap.none} ${DESIGN_SYSTEM.gap["4xs"]}`
+      : DESIGN_SYSTEM.gap.none};
 
-  border-radius: ${({ theme }) => theme.radius?.circle};
+  border-radius: ${DESIGN_SYSTEM.radius.circle};
 
-  background-color: ${({ theme }) =>
-    theme.colors?.light["feedback-notification"]};
-  color: ${({ theme }) => theme.colors?.light["object-static-inv-hero"]};
+  background-color: ${({ theme }) => theme.light["feedback-notification"]};
+  color: ${({ theme }) => theme.light["object-static-inv-hero"]};
 
   text-align: center;
-  ${({ theme }) => theme.typography?.body["2xs"]};
+  ${DESIGN_SYSTEM.typography.body["2xs"]};
 `;
